ZI.match {WormTools}	R Documentation

Matching two datafiles to eachother.

Description

If you have two datafiles with overlapping information, this function allows you to select the columns in which the identifier resides for both dataframes and produces a vector in which you put the values of either of the data.frame which you'd like to convert to the other data.frame. Therefore, this method also works if the nrow() differs for both data.frames.

Usage


ZI.match(data1, data2, value, type = "numeric")

Arguments

data1
A vector or data.frame column in which the identifier of data.frame 1 is.

data2
A vector or data.frame column in which the identifier of data.frame 2 is.

value
A vector or data.frame column in which you can find the values you'd like to return in an output vector.

type
Default: numeric, other options: 'character', make sure to select the option character if it is a character.

Examples


ZI.match

[Package WormTools version 0.0.0.9000 Index]