ZI.plot {WormTools}	R Documentation

Plot in Nathan's favourite style (hint: There are options!).

Description

This function elegantly puts together a graph using ggplot. Be sure ggplot2 and ggthemes are installed.

Usage


ZI.plot(data, x, y, group, fill, colour = fill, facet_wrap = F,
  facet = "", ylab = "y", xlab = "x", filling = "automatic",
  colours, type = "Boxplot")

Arguments

data
Here goes your data as a data-frame.

x
x-values (format: either data$column or column)

y
y-values (format: idem)

group
How you'd like to group your variables.

fill
How you'd like to fill the plots. Different from colour.

colour
Default == fill. How you'd like to colour the outlines of your plotted elements (e.g. outline of boxes in the boxplot)

facet_wrap
(T or F) Would you like to have several panels?

facet
Optional: If you chose for several panels, here you indicate what the panels should include.

ylab
Labels the y-axis

xlab
Labels the x-axis

filling
'manual' or 'automatic' (Default: automatic) How would you like to fill the plots? Manually or automatically?

colours
If you opted filling = 'manual', provide a vector of the colours you'd like. This should have the same length as the number of objects you graph.

type
'Boxplot' or 'violin' for now.

colour
How you'd like to colour your

Examples


ZI.plot

[Package WormTools version 0.0.0.9000 Index]