analyse.MSA {WormTools}	R Documentation

Analyse MSA data

Description

Analyse MSA data which you first fed into the python script "analyse_alignment_on_similarity_in_speciesv2.py"

Usage


analyse.MSA(folder = getwd(), destination = getwd(), writeCsv = T)

Arguments

folder
folder in which you saved the python output

destination
folder in which you'd like to save the output files

writeCsv
if you want to save the output files in a csv.

Examples


analyse.MSA

[Package WormTools version 0.0.0.9000 Index]