analyse.microbeta.data {WormTools}	R Documentation

Analyse your micro.beta data.

Description

Using this function, you can elegantly and rather quickly scout your microbeta screen for potential hits. The outcome will be a list of several elements, including a density plot, the potential hits and a levelplot.

Usage


analyse.microbeta.data(data, pool.plates = T, screen.type = "hit",
  pos.control = "ATP", baseline = "BSA", norm.over.ATP = T,
  set.to.zero = F, nr.plates = 5, plate.type = 96,
  graph.levelplots = F, path = path,
  levelplot.file.names = "Default", estimate.p.values = F,
  plot.levelplots = F, p.value.distribution = c("chisq", "gamma",
  "exp", "normal"), alpha = 0.05, plot.distr.estimate = F,
  graph.density = F)

Arguments

data
Data that were loaded using load.microbeta and subsequently summarised using process.microbeta.data()

pool.plates
would you like to pool the plates in your assay? IF not, you'll get one plot as output, otherwise it produces a list of several plots.

screen.type
Default= wide screen.

pos.control
= "ATP", your positive control.

baseline
= "BSA", the solution in which peptides were diluted.

norm.over.ATP
Default = F: Would you like to divide your values by the ATP value per plate or not?

set.to.zero
Default = T: would you like to set negative z.scores to 0?

nr.plates
The number of plates you screened.

plate.type
The type of plates you used (i.e. e.g. 96 wells)

graph.levelplots
Default: F; If T: writes a levelplot graph to the specified location in path

path
If you opted to write graphs to ppt: where should the file go?

levelplot.file.names
Default: 'Levelplot Plate 1', 2, 3, ... If graph2ppt: how should you like to call your files?

estimate.p.values
Default = F: Would you like to estimate p.values?

plot.levelplots
If T: levelplots are plotted, this can take a while

p.value.distribution
No Default: what distribution do you think you'r data has?

alpha
What alpha value would you like to identify hits on?

plot.distr.estimate
Default = F; Would you like to plot the estimated distribution of your data?

graph.density
Default = F; Woudl you like to write the density curve to a file?

Examples



analyse.microbeta.data

[Package WormTools version 0.0.0.9000 Index]