asterisks {WormTools}	R Documentation

Symbolizing significance by asterisks

Description

Adapted from prof. Tom Wenseleers: This function uses given threshold values (cutpoints) to code these in symbols (symbols) By default, it will annote asterisks (*) to p.value values.

Usage


asterisks(value, cutpoints = c(0, 0.001, 0.01, 0.05, 0.1, 1),
  symbols = c("***", "**", "*", ".", ""))

Arguments

value
Here goes your data.

cutpoints
Here goes a vector of cutpoints, usually one longer than the vector of thresholds

symbols
Here goes a vector of symbols, one shorter than the vector of cutpoints

Examples


asterisks

[Package WormTools version 0.0.0.9000 Index]