binning {WormTools}	R Documentation

Replace all NA values by a constant.

Description

This function enables you to replace all NA values by a constant.

Usage


binning(data, binrange, set_bins = "automatic", manual_bins = c())

Arguments

data
Here you put the data for which you like to replace it by bins.

binrange
Here you put the range that the bins should span. E.g. if you want to put bins for years, you can chose 5 years and that'll give you bins of five years wide.

set_bins
Optinoal argument: 'automatic', or 'manual'; Default: 'automatic'. If you like to set bins manually, select manual.

manual_bins
Here goes a vector with the bins you'd like. If you already selected automatic, this argument is optional.

Examples


binning

[Package WormTools version 0.0.0.9000 Index]