check.learningData {WormTools}	R Documentation

Checking your learning data whether it worked or not.

Description

This function creates a data.frame in which you can check the quality of your learning assay.

Usage


check.learningData(data, column, strain = "N2", learning.phenotype = T,
  threshold = 0.7, learning.cutpoints = c(-1, (threshold - 0.2),
  (threshold - 0.1), (threshold - 0.05), (threshold), 1),
  non.learning.cutpoints = c(-1, (threshold - 0.15), (threshold),
  (threshold + 0.1), (threshold + 0.2), 1))

Arguments

data
Here goes your data, loaded by load.learningdata()

column
data$column: Here goes the column you'd like to check: data$CI.

strain
what strain would you like to check up on? Default: N2

learning.phenotype
Is the strain learning (true) or shouldn't it be (false)? Default: True (learning)

threshold
The threshold CI value for which the assay went well. For naive: CI > threshold is good, whereas for conditioned CI < threshold is better

learning.cutpoints
The cutpoints you used to determine the significance of the learning phenotype

nonlearning.cutpoints
The cutpoints you used to determine the significance of the non-learning phenotype

Examples


check.learningData

[Package WormTools version 0.0.0.9000 Index]