folderTag {WormTools}	R Documentation

Produce a tag for every subfolder in a folder

Description

As the title states: this function will go over all subfolders in a folder and produce an identifier for each folder

Usage


folderTag(folder = getwd(), suffix = "")

Arguments

folder
This can be any folder. By default the function will check all subfolers the wd

Examples


folderTag

[Package WormTools version 0.0.0.9000 Index]