levelplot.data.microbeta {WormTools}	R Documentation

Produce data for a levelplot for your receptor screen data

Description

Returns a data.frame for receptor data to produce a levelplot after you processed them with process.microbeta.data

Usage


levelplot.data.microbeta(data, nr.plates = 5, plate.type = 96,
  nr.x = 12, nr.y = 8)

Arguments

data
This is the data you loaded with load.microbeta() and processed with process.microbeta.data()

pos.control
Default: 'ATP', this is the character string coding for the positive control (I wasn't entirely sure whether this is always ATP)

baseline
Default: 'BSA'. Idem, but for teh baseline values, i.e. the solvent in which you diluted your peptides.

norm.over.ATP
Default: True, this is an optional argument that you can tick if you'd like to normalize over the positive control.

Examples


levelplot.data.microbeta

[Package WormTools version 0.0.0.9000 Index]