levelplot.microbeta {WormTools}	R Documentation

Make a levelplot (i.e. a heatmap without clustering) for your microbeta data

Description

This function will allow you to plot your microbeta data to visually see whether you have potential hits.

Usage


levelplot.microbeta(data, pool.plates = F, nr.plates = 5,
  plate.type = 96, graph2ppt = F, graph.names = "Default",
  path = getwd())

Arguments

data
Data that were loaded using load.microbeta and subsequently summarised using process.microbeta.data()

pool.plates
would you like to pool the plates in your assay? IF not, you'll get one plot as output, otherwise it produces a list of several plots.

nr.plates
The number of plates you screened.

plate.type
The type of plates you used (i.e. e.g. 96 wells)

graph2ppt
Default: F; If T: writes a graph to the specified location in path

path
If graph2ppt: where should the file go?

file.names
Default: 'Levelplot Plate 1', 2, 3, ... If graph2ppt: how should you like to call your files?

Examples


levelplot.microbeta

[Package WormTools version 0.0.0.9000 Index]