load.LocomotionData {WormTools}	R Documentation

Loading data from the Locomotion Assay

Description

This function allows you to quickly load all data from one mother folder in which you saved data for several experiments in subfolders. E.g. the main folder contains your data, the subfolders contain .csv files for each time a different strain in a different condition.

Usage


load.LocomotionData(path = getwd(), folders, strains, conditions)

Arguments

path
Enter the path to the main folder.

folders
Enter all subfolders you constituted. I'll try to find a way to do this more efficiently.

strains
Enter a list of the strains you used. These should be ordered in the way you ordered the respective folders.

conditions
Enter a list of again respective conditions for all folders. Mind that the order does matter a lot!

Examples


load.LocomotionData

[Package WormTools version 0.0.0.9000 Index]