load.MultipleCsv {WormTools}	R Documentation

Loading multiple .csv files from a common folder.

Description

This function allows you to quickly load all data from one mother folder in which you saved data as .csv files.

Usage


load.MultipleCsv(add.replicate = F)

Arguments

NONE
Be sure to set your working directory to the folder of which you want to access the data. Your data should be formatted as follows: One column in which there is the data for the DA spot, one column in which there is the data for the EtOH spot and another column in which there is the data for all other worms. Next to this, the function requires a column that indicates the strain and a column that indicates the condition.

Examples


load.MultipleCsv

[Package WormTools version 0.0.0.9000 Index]