load.layout {WormTools}	R Documentation

Loading the plate layout.

Description

This function loads the plate layout and adds the relevant columns to the dataset. Load your .csv file as a data.frame with a first column in which the y-values (A, B, ..., H) are and the next few columns show the peptide per x-value (1, 2, ..., 12). All plates are pasted underneath eachother.

Usage


load.layout(filename, nr.plates = 5)

Arguments

filename
The name of the file.

nr.plates
Default == 5: the number of plates you screened.

Examples


load.layout

[Package WormTools version 0.0.0.9000 Index]