load.microbeta {WormTools}	R Documentation

Loading the data from the microbeta machine.

Description

This function loads your .csv file from the microbeta machine with an option to immediately read in the peptides from the plate layout file.

Usage


load.microbeta(data, layout, nr.plates = 5, nr.series = 30,
  matching = "automatic", filename = "microbeta.compiled.data.csv",
  write.csv = T)

Arguments

data
The data from your receptor screen

layout
Optional argument: The plate layout.

nr.plates
The number of plates you screened.

nr.series
The number of seconds in the peptide addition and triton addition measurements. Default: 30 (on microbeta)

matching
== c('manual', 'automatic') Default: 'automatic' Optional argument that by default loads in the plate layout as well. THIS CAN TAKE A WHILE.

filename
The name you'd like to give to the .csv file that you'll write.

write.csv
== c(T, F), Default: T. Optional argument to write a .csv file with your data.

Examples


load.microbeta

[Package WormTools version 0.0.0.9000 Index]