make.a.range {WormTools}	R Documentation

Make a range

Description

Make a range out of a vector of elements. This allows you to visually make it easier to find gropus of subsequent numbers.

Usage


make.a.range(values)

Arguments

values
the vector of values in which you see several ranges

Examples


make.a.range

[Package WormTools version 0.0.0.9000 Index]