na.col.omit {WormTools}	R Documentation

Omitting NA values (or rows) for a specific column of your dataset

Description

Removing NA values from a specific column of your dataset: If it's a data.frame, the row is omitted, if it is a

Usage


na.col.omit(data, column)

Arguments

data
Here goes your data. Accepted datatypes e.g. data.frame, lists, ...

column
data$column: Here goes the column you'd like to check.

Examples


NaN.omit

[Package WormTools version 0.0.0.9000 Index]