na.replace {WormTools}	R Documentation

Replacing NA values by a constant.

Description

This function allows you to replace any NA that occur by a value. Chose the value wisely so that it will not interfere with your data, but will allow you to e.g. divide it or execute logical statements.

Usage


na.replace(data, replacement)

Arguments

data
Here goes the data

replacement
Here goes your replacement

Examples


na.replace

[Package WormTools version 0.0.0.9000 Index]