plot.blank {WormTools}	R Documentation

Plot a blank screen

Description

Sometimes the plotting window does a little onorthodox things and you might want to plot a blank screen in between.

Usage


plot.blank()

Examples


plot.blank

[Package WormTools version 0.0.0.9000 Index]