set.to.zero {WormTools}	R Documentation

Set a series of values to 0 (or a threshold value)

Description

This function will set any value to an arbitrary value if below a threshold value (for now). If you enter a column in a dataset, mind to indeed enter the column!

Usage


set.to.zero(value, output = 0, threshold = 0)

Arguments

value
The values you want to set to a threshold if

output
Choose the output value you'd like.

threshold
Choose the threshold

Examples


set.to.zero

[Package WormTools version 0.0.0.9000 Index]