string.substringRight {WormTools}	R Documentation

Generate a tail of a string!

Description

Generate a substring composed of the final n characters of your string in your string of interest

Usage


string.substringRight(x, n)

Arguments

x
A string you'd like to slice

n
the number of characters your string should contain.

Examples


string.substringRight

[Package WormTools version 0.0.0.9000 Index]