z.score {WormTools}	R Documentation

Calculate the z-score.

Description

Calculate the z-score of a value by manually indicating the mean and the sd/var.

Usage


z.score(value, mean, sd, var, method = "sd")

Arguments

value
The value of which you'd like to calculate the z-score.

mean
The mean value

sd
The sd of the distribution of your value.

var
The variance of the distribution. By default, the sd is used, but if you like you can also put the variance in;

method=='sd'
By default you put the sd in, however, the variance can also work, if you indicate that you're using the variance;

Examples


z.score

[Package WormTools version 0.0.0.9000 Index]